+++ flutter_quill_extensions/lib/models/config/toolbar/buttons/media_button.dart
@@ -38,7 +38,6 @@ class QuillToolbarMediaButtonOptions extends QuillToolbarBaseButtonOptions<
     super.tooltip,
     super.iconTheme,
     super.childBuilder,
-    super.controller,
   });
 
   final double? iconSize;