+++ lib/src/models/config/toolbar/buttons/select_header_style_configurations.dart
@@ -19,7 +19,6 @@ class QuillToolbarSelectHeaderStyleButtonsOptions
   const QuillToolbarSelectHeaderStyleButtonsOptions({
     super.afterButtonPressed,
     super.childBuilder,
-    super.controller,
     super.iconTheme,
     super.tooltip,
     this.axis,